+++ Core/GoalsComponent/WaitResult.cs
@@ -3,15 +3,21 @@
 public readonly struct WaitResult
 {
     private readonly bool timeout;
-    private readonly double elapsedMs;
+    private readonly float elapsedMs;
 
-    public WaitResult(bool timeout, double elapsedMs)
+    // Timeout = true
+    // elapsed = negative
+
+    // Timeout = false
+    // elapsed = positive
+
+    public WaitResult(bool timeout, float elapsedMs)
     {
         this.timeout = timeout;
         this.elapsedMs = elapsedMs;
     }
 
-    public void Deconstruct(out bool timeout, out double elapsedMs)
+    public void Deconstruct(out bool timeout, out float elapsedMs)
     {
         timeout = this.timeout;
         elapsedMs = this.elapsedMs;
